refactor!: rename the `Options` class
