+++ source/config/index.ts
@@ -1,7 +1,7 @@
 export { ConfigDiagnosticText } from "./ConfigDiagnosticText.js";
 export { Config, type ResolvedConfig } from "./Config.js";
 export { OptionBrand } from "./OptionBrand.enum.js";
-export { type ItemDefinition, type OptionDefinition, OptionDefinitionsMap } from "./OptionDefinitionsMap.js";
+export { type ItemDefinition, type OptionDefinition, Options } from "./Options.js";
 export { OptionGroup } from "./OptionGroup.enum.js";
 export { defaultOptions } from "./defaultOptions.js";
 export type { CommandLineOptions, ConfigFileOptions } from "./types.js";
